Created the LearnLodash course for 42.
The idea is very simple we create a whole bunch of self contained labs.
These labs will contain imperative code which needs to be rewritten to
a more functional approach with lodash.

Each lab will have unit tests which verify that the imperative code
works. These test will then prove that the new lodash code works as well.
The student should never change the unit tests and only needs to write
lodash code.

The projects setup is really easy. There is a folder called 'course' which
will contain all labs. Then there is a karma.conf.js which loads in all
courses and runs the unit tests.

In the README there are instructions which detail how to setup the dev
environment and how to start the course.